requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.0
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
python-dotenv==1.0.0